# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "같이살림 API",
    "version": "1.0.0",
    "description": "같이살림 API 명세서입니다."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "알림"
        ],
        "summary": "알림 목록 조회",
        "description": "인증된 사용자의 숨기지 않은 알림을 최신순으로 조회합니다.",
        "operationId": "getNotifications",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "data",
                    "error"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/NotificationListResponse"
                    },
                    "error": {
                      "nullable": true,
                      "example": null
                    }
                  }
                },
                "example": {
                  "statusCode": 200,
                  "data": {
                    "notifications": [
                      {
                        "notificationId": 101,
                        "groupId": 3,
                        "type": "CHORE_DUE",
                        "refId": 42,
                        "message": "오늘까지 완료할 집안일이 있습니다.",
                        "isRead": false,
                        "createdAt": "2026-07-16T09:00:00.000Z"
                      }
                    ]
                  },
                  "error": null
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/notifications/preferences": {
      "get": {
        "tags": [
          "알림"
        ],
        "summary": "알림 설정 조회",
        "description": "사용자별 알림 설정을 조회합니다. 저장된 설정이 없으면 모든 항목을 true로 반환합니다.",
        "operationId": "getNotificationPreferences",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "알림 설정 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "data",
                    "error"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/NotificationPreferences"
                    },
                    "error": {
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "patch": {
        "tags": [
          "알림"
        ],
        "summary": "알림 설정 수정",
        "description": "전달된 설정만 수정합니다. 비활성화한 유형도 앱 내 알림은 유지하고 웹 푸시만 중단합니다. groupActivity는 그룹에 속한 모든 알림의 상위 설정입니다.",
        "operationId": "patchNotificationPreferences",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationPreferences"
              },
              "example": {
                "choreDue": true,
                "newMessage": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "알림 설정 수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "data",
                    "error"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/NotificationPreferences"
                    },
                    "error": {
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/notifications/unread-count": {
      "get": {
        "tags": [
          "알림"
        ],
        "summary": "미읽음 알림 개수 조회",
        "description": "인증된 사용자의 숨기지 않은 미읽음 알림 개수를 조회합니다.",
        "operationId": "getNotificationsUnreadCount",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "data",
                    "error"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/UnreadNotificationCountResponse"
                    },
                    "error": {
                      "nullable": true,
                      "example": null
                    }
                  }
                },
                "example": {
                  "statusCode": 200,
                  "data": {
                    "unreadCount": 4
                  },
                  "error": null
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/notifications/{notificationId}/read": {
      "patch": {
        "tags": [
          "알림"
        ],
        "summary": "단일 알림 읽음 처리",
        "description": "인증된 사용자의 노출 중인 알림을 읽음 처리합니다.",
        "operationId": "patchNotificationsNotificationIdRead",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "description": "알림 ID",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "data",
                    "error"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/ReadNotificationResponse"
                    },
                    "error": {
                      "nullable": true,
                      "example": null
                    }
                  }
                },
                "example": {
                  "statusCode": 200,
                  "data": {
                    "notificationId": 101,
                    "isRead": true
                  },
                  "error": null
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/notifications/read-all": {
      "patch": {
        "tags": [
          "알림"
        ],
        "summary": "전체 읽음 처리",
        "description": "인증된 사용자의 노출 중인 미읽음 알림을 모두 읽음 처리합니다.",
        "operationId": "patchNotificationsReadAll",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "data",
                    "error"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/ReadAllNotificationsResponse"
                    },
                    "error": {
                      "nullable": true,
                      "example": null
                    }
                  }
                },
                "example": {
                  "statusCode": 200,
                  "data": {
                    "updatedCount": 3
                  },
                  "error": null
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/notifications/{notificationId}": {
      "delete": {
        "tags": [
          "알림"
        ],
        "summary": "알림 숨김 처리",
        "description": "인증된 사용자의 알림을 soft hide 처리합니다.",
        "operationId": "deleteNotificationsNotificationId",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "description": "알림 ID",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "알림 숨김 처리 성공"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Cognito access token을 Authorization: Bearer 헤더로 전달합니다."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "data": {
            "nullable": true
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "errors": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
...
      "NotificationPreferences": {
        "type": "object",
        "required": [
          "choreDue",
          "supplyStatusChanged",
          "newMessage",
          "expenseRequest",
          "ruleAgreementRequest",
          "groupActivity"
        ],
        "properties": {
          "choreDue": {
            "type": "boolean",
            "example": true
          },
          "supplyStatusChanged": {
            "type": "boolean",
            "example": true
          },
          "newMessage": {
            "type": "boolean",
            "example": true
          },
          "expenseRequest": {
            "type": "boolean",
            "example": true
          },
          "ruleAgreementRequest": {
            "type": "boolean",
            "example": true
          },
          "groupActivity": {
            "type": "boolean",
            "description": "그룹에 속한 모든 알림의 웹 푸시 상위 설정",
            "example": true
          }
        }
      },
      "UpdateNotificationPreferences": {
        "type": "object",
        "minProperties": 1,
        "properties": {
          "choreDue": {
            "type": "boolean"
          },
          "supplyStatusChanged": {
            "type": "boolean"
          },
          "newMessage": {
            "type": "boolean"
          },
          "expenseRequest": {
            "type": "boolean"
          },
          "ruleAgreementRequest": {
            "type": "boolean"
          },
          "groupActivity": {
            "type": "boolean",
            "description": "그룹에 속한 모든 알림의 웹 푸시 상위 설정"
          }
        }
      },
      "Notification": {
        "type": "object",
        "required": [
          "notificationId",
          "groupId",
          "type",
          "refId",
          "message",
          "isRead",
          "createdAt"
        ],
        "properties": {
          "notificationId": {
            "type": "integer",
            "example": 101
          },
          "groupId": {
            "type": "integer",
            "nullable": true,
            "example": 3
          },
          "type": {
            "type": "string",
            "enum": [
              "CHORE_DUE",
              "EXPENSE_REQUEST",
              "SUPPLY_LOW",
              "RULE_CHANGED",
              "NEW_MESSAGE",
              "GROUP_INVITE",
              "EXPENSE_DONE"
            ],
            "example": "CHORE_DUE"
          },
          "refId": {
            "type": "integer",
            "nullable": true,
            "example": 42
          },
          "message": {
            "type": "string",
            "example": "오늘까지 완료할 집안일이 있습니다."
          },
          "isRead": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-16T09:00:00.000Z"
          }
        }
      },
      "NotificationListResponse": {
        "type": "object",
        "required": [
          "notifications"
        ],
        "properties": {
          "notifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notification"
            }
          }
        }
      },
      "UnreadNotificationCountResponse": {
        "type": "object",
        "required": [
          "unreadCount"
        ],
        "properties": {
          "unreadCount": {
            "type": "integer",
            "example": 4
          }
        }
      },
...
      "ReadNotificationResponse": {
        "type": "object",
        "required": [
          "notificationId",
          "isRead"
        ],
        "properties": {
          "notificationId": {
            "type": "integer",
            "example": 101
          },
          "isRead": {
            "type": "boolean",
            "enum": [
              true
            ],
            "example": true
          }
        }
      },
      "ReadAllNotificationsResponse": {
        "type": "object",
        "required": [
          "updatedCount"
        ],
        "properties": {
          "updatedCount": {
            "type": "integer",
            "example": 3
          }
        }
      }
    },
    "responses": {
      "ErrorResponse": {
        "description": "오류 응답",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}